Initial document for tool-conventions
This is an initial proposal of a document for the tool-conventions
repository.

It focuses more on the format itself that on the goals and design.

This is still incomplete and the format can still totally change,
but having an entry in tool-conventions is important to get tool
maintainers aware of our effort.

Posted here for review and discussions before we are ready to submit
it to the proper repos.